배포 이후 작업

+++ app/Services/Helper.php
@@ -212,7 +212,6 @@ EEE;
             echo "<script type=\"text/javascript\">
                     $(document).ready(function(){
                         jalert_url('".$msg."', 'back', '알림');
-
                     });
                 </script>";
         } else {
@@ -456,60 +455,40 @@ EEE;
     public function rotateImage($file)
     {
         // 회전 된 이미지 그대로 보여주기
-        $exif = exif_read_data($file);
+        if (exif_imagetype($file) && exif_imagetype($file) === IMAGETYPE_JPEG) {
+            $exif = exif_read_data($file);
 
-        if(isset($exif['Orientation'])) {
-            $orientation = $exif['Orientation'];
+            if(isset($exif['Orientation'])) {
+                $orientation = $exif['Orientation'];
 
-            if($orientation != 1){
-                $img = imagecreatefromjpeg($file);
+                if($orientation != 1){
+                    $img = imagecreatefromjpeg($file);
 
-                if (in_array($exif['Orientation'], [3, 4])) {
-                    $img = imagerotate($img, 180, 0);
+                    if (in_array($exif['Orientation'], [3, 4])) {
+                        $img = imagerotate($img, 180, 0);
+                    }
+                    if (in_array($exif['Orientation'], [5, 6])) {
+                        $img = imagerotate($img, -90, 0);
+                    }
+                    if (in_array($exif['Orientation'], [7, 8])) {
+                        $img = imagerotate($img, 90, 0);
+                    }
+                    if (in_array($exif['Orientation'], [2, 5, 7, 4])) {
+                        imageflip($img, IMG_FLIP_HORIZONTAL);
+                    }
+
+                    $tempPath = tempnam(sys_get_temp_dir(), 'rotated_image');
+                    imagejpeg($img, $tempPath);
+
+                    // 메모리 해제
+                    imagedestroy($img);
+
+                    // 임시 파일을 원본 파일로 복사 또는 이동
+                    copy($tempPath, $file);
+
+                    // 임시 파일 삭제
+                    unlink($tempPath);
                 }
-                if (in_array($exif['Orientation'], [5, 6])) {
-                    $img = imagerotate($img, -90, 0);
-                }
-                if (in_array($exif['Orientation'], [7, 8])) {
-                    $img = imagerotate($img, 90, 0);
-                }
-                if (in_array($exif['Orientation'], [2, 5, 7, 4])) {
-                    imageflip($img, IMG_FLIP_HORIZONTAL);
-                }
-
-                // 이미지를 회전시키기
-//                switch($orientation){
-//                    case 3:
-//                    case 4:
-//                        $img = imagerotate($img, 180, 0);
-//                        break;
-//                    case 5:
-//                    case 6:
-//                        $img = imagerotate($img, -90, 0);
-//                        break;
-//                    case 7:
-//                    case 8:
-//                        $img = imagerotate($img, 90, 0);
-//                        break;
-//                    case 2:
-//                    case 4:
-//                    case 5:
-//                    case 7:
-//                        imageflip($img, IMG_FLIP_HORIZONTAL);
-//                        break;
-//                }
-
-                $tempPath = tempnam(sys_get_temp_dir(), 'rotated_image');
-                imagejpeg($img, $tempPath);
-
-                // 메모리 해제
-                imagedestroy($img);
-
-                // 임시 파일을 원본 파일로 복사 또는 이동
-                copy($tempPath, $file);
-
-                // 임시 파일 삭제
-                unlink($tempPath);
             }
         }
